# Some performance stuff and some cleanup