--- a/cmd/really/really.cpp
+++ b/cmd/really/really.cpp
@@ -153,7 +153,7 @@ public:
     server->publishIntentResponse(quicr_namespace, result);
   };
 
-  virtual void onPublishIntentEnd(const quicr::Namespace& quicr_namespace,
+  virtual void onPublishIntentEnd(const quicr::Namespace& /* quicr_namespace */,
                                   const std::string& /* auth_token */,
                                   quicr::bytes&& /* e2e_token */)
   {