--- a/include/quicr/hex_endec.h
+++ b/include/quicr/hex_endec.h
@@ -112,10 +112,7 @@ public:
       bits |= (value & ~(~0x0ull << dist));
     };
 
-    char hex[Size / 4 + 1];
-    std::snprintf(hex, Size / 4 + 1, "%0*llX", Size / 4, bits);
-
-    return std::string("0x") + hex;
+    return std::string("0x") + uint_to_hex<uint64_t>(bits);
   }
 
   template<bool B = Size <= sizeof(uint64_t) * 8>
@@ -160,14 +157,8 @@ public:
 
     std::string out_hex = "0x";
     for (size_t i = 0; i < Size / sizeof_uint64_bits; ++i) {
-      char hex[sizeof(uint64_t) * 2 + 1];
-      std::snprintf(
-        hex,
-        sizeof(uint64_t) * 2 + 1,
-        "%0*llX",
-        int(sizeof(uint64_t) * 2),
+      out_hex += uint_to_hex<uint64_t>(
         ((bits & bits_mask) >> (Size - sizeof_uint64_bits)).to_ullong());
-      out_hex += hex;
       bits <<= sizeof_uint64_bits;
     }
 
@@ -207,7 +198,7 @@ public:
     typename Uint_t = uint64_t,
     typename = typename std::enable_if<is_valid_uint<Uint_t>::value, Uint_t>>
   static constexpr std::array<Uint_t, sizeof...(Dist)> Decode(
-    const std::string_view& hex)
+    std::string_view hex)
   {
     static_assert((Size & (Size - 1)) == 0, "Size must be a power of 2");
     static_assert(Size >= (Dist + ...),
@@ -247,13 +238,13 @@ public:
     typename = typename std::enable_if<is_valid_uint<Uint_t>::value, Uint_t>>
   static constexpr typename std::enable_if<B, std::vector<Uint_t>>::type Decode(
     std::span<uint8_t> distribution,
-    const std::string_view& hex)
+    std::string_view hex)
   {
     static_assert((Size & (Size - 1)) == 0, "Size must be a power of 2");
 
     const auto dist_size = distribution.size();
     std::vector<uint64_t> result(dist_size);
-    uint64_t bits = hex_to_uint(hex);
+    uint64_t bits = hex_to_uint<uint64_t>(hex);
     for (int i = dist_size - 1; i >= 0; --i) {
       const auto dist = distribution[i];
       result[i] = bits & ~(~0x0ull << dist);
@@ -268,7 +259,7 @@ public:
     bool B = Size <= sizeof(uint64_t) * 8,
     typename = typename std::enable_if<is_valid_uint<Uint_t>::value, Uint_t>>
   static constexpr typename std::enable_if<!B, std::vector<Uint_t>>::type
-  Decode(std::span<uint8_t> distribution, const std::string_view& hex)
+  Decode(std::span<uint8_t> distribution, std::string_view hex)
   {
     static_assert((Size & (Size - 1)) == 0, "Size must be a power of 2");
 
@@ -289,8 +280,9 @@ public:
     constexpr size_t num_sections = Size / sizeof_uint64_bits;
     for (size_t i = start_pos, j = 0; i < (section_length * num_sections);
          i += section_length) {
-      bits |= std::bitset<Size>(hex_to_uint(hex.substr(i, section_length)))
-              << (sizeof_uint64_bits * (num_sections - ++j));
+      bits |=
+        std::bitset<Size>(hex_to_uint<uint64_t>(hex.substr(i, section_length)))
+        << (sizeof_uint64_bits * (num_sections - ++j));
     }
 
     const auto dist_size = distribution.size();