--- a/include/quicr/quicr_namespace.h
+++ b/include/quicr/quicr_namespace.h
@@ -2,11 +2,13 @@
 
 #include <quicr/quicr_name.h>
 
-namespace quicr {
-namespace messages {
-class MessageBuffer;
-}
+#include <ostream>
 
+namespace quicr {
+
+/**
+ * @brief A prefix for a quicr::Name
+ */
 class Namespace
 {
 public:
@@ -20,20 +22,27 @@ public:
   Namespace& operator=(Namespace&& other) = default;
 
   bool contains(const Name& name) const;
-  bool contains(const Namespace& name_space) const;
-  constexpr Name name() const { return _mask_name; }
+  bool contains(const Namespace& prefix) const { return contains(prefix._name); }
+
+  constexpr Name name() const { return _name; }
   constexpr uint8_t length() const { return _sig_bits; }
   std::string to_hex() const;
 
   friend bool operator==(const Namespace& a, const Namespace& b);
   friend bool operator!=(const Namespace& a, const Namespace& b);
+
   friend bool operator>(const Namespace& a, const Namespace& b);
+  friend bool operator>(const Namespace& a, const Name& b);
+  friend bool operator>(const Name& a, const Namespace& b);
+
   friend bool operator<(const Namespace& a, const Namespace& b);
+  friend bool operator<(const Namespace& a, const Name& b);
+  friend bool operator<(const Name& a, const Namespace& b);
 
   friend std::ostream& operator<<(std::ostream& os, const Namespace& ns);
 
 private:
-  Name _mask_name;
+  Name _name;
   uint8_t _sig_bits;
 };
 }