… benchmark/name.cpp
… src/quicr_namespace.cpp
… test/encode.cpp
… test/hex_endec.cpp
… test/name.cpp
… test/namespace.cpp
… test/quicr_client.cpp