Weekly attempt to improve benchmark numbers. This PRs numbers:
```
Run on (8 X 24 MHz CPU s)
CPU Caches:
  L1 Data 64 KiB
  L1 Instruction 128 KiB
  L2 Unified 4096 KiB (x8)
Load Average: 2.32, 1.94, 1.57
----------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations
----------------------------------------------------------------------------
BM_NameConstructFromHexString           29.4 ns         29.4 ns     23692833
BM_NameConstructFromHexStringView       28.5 ns         28.5 ns     24538587
BM_NameConstructFromVector              1.47 ns         1.47 ns    475617794
BM_NameConstructFromBytePointer         1.18 ns         1.18 ns    594479830
BM_NameCopyConstruct                   0.000 ns        0.000 ns   1000000000
BM_NameLeftShift                       0.859 ns        0.859 ns    792796874
BM_NameRightShift                      0.861 ns        0.861 ns    812998688
BM_NameAdd                              1.91 ns         1.91 ns    366935928
BM_NameSub                              1.98 ns         1.98 ns    352746129
BM_NameToHex                            45.8 ns         45.8 ns     14794525
BM_NameRealArithmetic                   19.3 ns         19.3 ns     35832019
BM_MessageBufferConstruct               30.4 ns         30.4 ns     23053009
BM_MessageBufferPushBack                6.84 ns         6.84 ns    100329655
BM_MessageBufferPushBack16              7.07 ns         7.07 ns    100860193
BM_MessageBufferPushBack32              7.49 ns         7.49 ns     97002619
BM_MessageBufferPushBack64              8.45 ns         8.45 ns     89314195
BM_MessageBufferPushBackVector           283 ns          282 ns      3043571
BM_MessageBufferWriteName               36.0 ns         35.9 ns     19847458
BM_HexEndecEncode4x32_to_128            69.1 ns         69.1 ns     10186708
BM_HexEndecDecode128_to_4x32            87.0 ns         87.0 ns      8041263
BM_HexEndecEncode4x16_to_64             14.3 ns         14.3 ns     49008626
BM_HexEndecDecode64_to_4x16             30.9 ns         30.9 ns     22480859
BM_HexEndecRealEncode                   80.5 ns         80.5 ns      8748141
BM_HexEndecRealDecode                    170 ns          170 ns      4158918
```
Vs `main`:
```
Run on (8 X 24 MHz CPU s)
CPU Caches:
  L1 Data 64 KiB
  L1 Instruction 128 KiB
  L2 Unified 4096 KiB (x8)
Load Average: 2.50, 1.73, 1.43
--------------------------------------------------------------------------
Benchmark                                Time             CPU   Iterations
--------------------------------------------------------------------------
BM_NameConstructFromHex               85.8 ns         85.8 ns      8132537
BM_NameConstructFromVector           0.881 ns        0.881 ns    799643587
BM_NameConstructFromBytePointer      0.863 ns        0.863 ns    815546649
BM_NameCopyConstruct                 0.000 ns        0.000 ns   1000000000
BM_NameLeftShift                      2.39 ns         2.39 ns    295603115
BM_NameRightShift                     2.30 ns         2.30 ns    303186490
BM_NameAdd                            1.91 ns         1.91 ns    363053592
BM_NameSub                            1.98 ns         1.98 ns    352671487
BM_NameToHex                           112 ns          112 ns      6318830
BM_NameRealArithmetic                 19.5 ns         19.5 ns     36125117
BM_MessageBufferConstruct             30.3 ns         30.3 ns     23137360
BM_MessageBufferPushBack              6.85 ns         6.85 ns     99383820
BM_MessageBufferPushBack16            7.09 ns         7.09 ns    100850022
BM_MessageBufferPushBack32            7.44 ns         7.44 ns     97759902
BM_MessageBufferPushBack64            8.40 ns         8.40 ns     89611470
BM_MessageBufferPushBackVector         289 ns          285 ns      3036402
BM_MessageBufferWriteName             36.1 ns         36.1 ns     19915955
BM_HexEndecEncode4x32_to_128           146 ns          146 ns      4844693
BM_HexEndecDecode128_to_4x32           156 ns          156 ns      4506883
BM_HexEndecEncode4x16_to_64           50.6 ns         50.6 ns     13956456
BM_HexEndecDecode64_to_4x16           31.8 ns         31.8 ns     21998674
BM_HexEndecRealEncode                  157 ns          157 ns      4466736
BM_HexEndecRealDecode                  288 ns          288 ns      2427344